tools/threat-intel-sync/requirements.txt
httpx>=0.27,<1.0
pyyaml>=6.0,<7.0
